+++ http_cache_client/CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix: Improve cache save time.
 - fix: Improve request date assignment.
 - fix: bump http to ^1.2.0
+- fix: `CacheOptions#maxStale` usage may lead to empty response body.
 
 ## 1.0.0
 - Initial version.