+++ http_cache_core/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.0
 - chore: Add writeContent to CacheResponse.
+- fix: `CacheOptions#maxStale` usage may lead to empty response body.
 
 ## 1.0.0
 - Initial version.